requirements.txt
telebot
pdfkit
beautifulsoup4
requests
python-dotenv